refactor(developer_view): developer view moved from examples to view module in applications
