+++ bec_widgets/applications/views/developer_view/developer_widget.py
