+++ tests/unit_tests/test_developer_view.py
@@ -17,7 +17,7 @@ from unittest import mock
 import pytest
 from qtpy.QtWidgets import QDialog
 
-from bec_widgets.examples.developer_view.developer_widget import DeveloperWidget
+from bec_widgets.applications.views.developer_view.developer_widget import DeveloperWidget
 from bec_widgets.widgets.editors.monaco.monaco_dock import MonacoDock
 from bec_widgets.widgets.editors.monaco.monaco_widget import MonacoWidget
 from bec_widgets.widgets.utility.ide_explorer.ide_explorer import IDEExplorer
